Move forms to namespaces and add more tests
