--- app/channels/application_cable/channel.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-##
-# Module regarding ActionCable (Rails implementation of WebSockets), generated automatically during new application creation.
-#
-module ApplicationCable
-  ##
-  # Base class for ActionCable Channel, generated automatically during new application creation
-  #
-  class Channel < ActionCable::Channel::Base
-  end
-end
