--- app/channels/application_cable/connection.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-##
-# Module regarding ActionCable (Rails implementation of WebSockets), generated automatically during new application creation.
-#
-module ApplicationCable
-  ##
-  # Base class for ActionCable Connection, generated automatically during new application creation
-  #
-  class Connection < ActionCable::Connection::Base
-  end
-end
