--- app/jobs/application_job.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-##
-# Base class for async jobs, generated automatically during new application creation
-#
-class ApplicationJob < ActiveJob::Base
-  # Automatically retry jobs that encountered a deadlock
-  # retry_on ActiveRecord::Deadlocked
-
-  # Most jobs are safe to ignore if the underlying records are no longer available
-  # discard_on ActiveJob::DeserializationError
-end
