+++ spec/forms/vehicles_management/submission_method_form_spec.rb
@@ -2,7 +2,7 @@
 
 require 'rails_helper'
 
-describe SubmissionMethodForm do
+describe VehiclesManagement::SubmissionMethodForm do
   subject { described_class.new(submission_method: submission_method) }
 
   let(:submission_method) { 'upload' }
