+++ spec/models/user_management/user_spec.rb
